Remove deprecated content and pgf modules, refactor Notebook class to integrate mime content handling directly. Update tests to ensure coverage for new functionality and remove obsolete tests.
