+++ tests/conftest.py
@@ -9,8 +9,7 @@ from nbstore.store import Store
 
 @pytest.fixture(scope="session")
 def src_dirs() -> list[Path]:
-    src_dir = Path(__file__).parent / "src"
-    return [src_dir / "jupyter", src_dir / "python"]
+    return [Path(__file__).parent / "src"]
 
 
 @pytest.fixture(scope="session")
