+++ tests/notebook/test_pdf.py
@@ -6,23 +6,21 @@ from nbstore.store import Store
 
 @pytest.fixture(scope="module")
 def nb(store: Store):
-    return store.get_notebook("svg.ipynb")
-
-
-def test_cell(nb: Notebook):
-    cell = nb.get_cell("fig:svg")
-    assert isinstance(cell, dict)
-    assert "cell_type" in cell
+    nb = store.read_notebook("pdf.ipynb")
+    assert not nb.is_executed
+    nb.execute()
+    assert nb.is_executed
+    return nb
 
 
 def test_source(nb: Notebook):
-    source = nb.get_source("fig:svg")
+    source = nb.get_source("fig:pdf")
     assert isinstance(source, str)
     assert "plot" in source
 
 
 def test_outputs(nb: Notebook):
-    outputs = nb.get_outputs("fig:svg")
+    outputs = nb.get_outputs("fig:pdf")
     assert isinstance(outputs, list)
     assert len(outputs) == 2
     assert isinstance(outputs[0], dict)
@@ -32,19 +30,18 @@ def test_outputs(nb: Notebook):
     assert outputs[1]["output_type"] == "display_data"
 
 
+def test_mime_content(nb: Notebook):
+    data = nb.get_mime_content("fig:pdf")
+    assert isinstance(data, tuple)
+    assert len(data) == 2
+    assert data[0] == "application/pdf"
+    assert isinstance(data[1], bytes)
+
+
 def test_data(nb: Notebook):
-    data = nb.get_data("fig:svg")
+    data = nb.get_data("fig:pdf")
     assert isinstance(data, dict)
     assert len(data) == 3
     assert "text/plain" in data
     assert "image/png" in data
-    assert data["image/svg+xml"].startswith('<?xml version="1.0"')
-
-
-def test_mime_content(nb: Notebook):
-    mime_content = nb.get_mime_content("fig:svg")
-    assert isinstance(mime_content, tuple)
-    mime, content = mime_content
-    assert mime == "image/svg+xml"
-    assert isinstance(content, str)
-    assert content.startswith('<?xml version="1.0"')
+    assert data["application/pdf"].startswith("JVBE")
