+++ tests/notebook/test_svg.py
@@ -6,27 +6,23 @@ from nbstore.store import Store
 
 @pytest.fixture(scope="module")
 def nb(store: Store):
-    nb = store.get_notebook("png.ipynb")
-    assert not nb.is_executed
-    nb.execute()
-    assert nb.is_executed
-    return nb
+    return store.read_notebook("svg.ipynb")
 
 
 def test_cell(nb: Notebook):
-    cell = nb.get_cell("fig:png")
+    cell = nb.get_cell("fig:svg")
     assert isinstance(cell, dict)
     assert "cell_type" in cell
 
 
 def test_source(nb: Notebook):
-    source = nb.get_source("fig:png")
+    source = nb.get_source("fig:svg")
     assert isinstance(source, str)
     assert "plot" in source
 
 
 def test_outputs(nb: Notebook):
-    outputs = nb.get_outputs("fig:png")
+    outputs = nb.get_outputs("fig:svg")
     assert isinstance(outputs, list)
     assert len(outputs) == 2
     assert isinstance(outputs[0], dict)
@@ -37,17 +33,18 @@ def test_outputs(nb: Notebook):
 
 
 def test_data(nb: Notebook):
-    data = nb.get_data("fig:png")
+    data = nb.get_data("fig:svg")
     assert isinstance(data, dict)
-    assert len(data) == 2
+    assert len(data) == 3
     assert "text/plain" in data
     assert "image/png" in data
-    assert data["image/png"].startswith("iVBO")
+    assert data["image/svg+xml"].startswith('<?xml version="1.0"')
 
 
 def test_mime_content(nb: Notebook):
-    data = nb.get_mime_content("fig:png")
-    assert isinstance(data, tuple)
-    assert len(data) == 2
-    assert data[0] == "image/png"
-    assert isinstance(data[1], bytes)
+    mime_content = nb.get_mime_content("fig:svg")
+    assert isinstance(mime_content, tuple)
+    mime, content = mime_content
+    assert mime == "image/svg+xml"
+    assert isinstance(content, str)
+    assert content.startswith('<?xml version="1.0"')
